requirements.txt
customtkinter
pillow
requests
geopy
dotenv